# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampaignForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List user's organizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{orgId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization by ID
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Workspaces ───────────────────────────────────────────────────────────
...
  /smtp-connections:
    get:
      operationId: listSmtpConnections
      tags: [smtp]
      summary: List SMTP connections
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: SMTP connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmtpConnection"
    post:
      operationId: createSmtpConnection
      tags: [smtp]
      summary: Create an SMTP connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmtpConnectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpConnection"
  /smtp-connections/{smtpId}:
    get:
      operationId: getSmtpConnection
      tags: [smtp]
      summary: Get SMTP connection
      parameters:
        - name: smtpId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SMTP connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpConnection"
    patch:
      operationId: updateSmtpConnection
      tags: [smtp]
      summary: Update SMTP connection
      parameters:
        - name: smtpId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmtpConnectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpConnection"
    delete:
      operationId: deleteSmtpConnection
      tags: [smtp]
      summary: Delete SMTP connection
      parameters:
        - name: smtpId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /smtp-connections/{smtpId}/test:
    post:
      operationId: testSmtpConnection
      tags: [smtp]
      summary: Test SMTP connection
      parameters:
        - name: smtpId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpTestResult"

  # ── API Keys ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Organization:
      type: object
      required: [id, name, slug, ownerId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        ownerId:
          type: integer
        plan:
          type: string
          enum: [free, starter, pro, business, enterprise]
        createdAt:
          type: string
    OrganizationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
        logoUrl:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        logoUrl:
          type: string

    # ── Workspaces ─────────────────────────────────────────────────────────
...
    SmtpConnection:
      type: object
      required: [id, workspaceId, name, provider, host, port, isDefault, isVerified, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        provider:
          type: string
          enum: [amazon_ses, sendgrid, mailgun, postmark, brevo, smtp2go, mailjet, resend, sparkpost, custom]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        tls:
          type: boolean
        isDefault:
          type: boolean
        isVerified:
          type: boolean
        dailySentCount:
          type: integer
        createdAt:
          type: string
    SmtpConnectionInput:
      type: object
      required: [workspaceId, name, provider, host, port, username, password]
      properties:
        workspaceId:
          type: integer
        name:
          type: string
          minLength: 1
        provider:
          type: string
          enum: [amazon_ses, sendgrid, mailgun, postmark, brevo, smtp2go, mailjet, resend, sparkpost, custom]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        tls:
          type: boolean
          default: true
        isDefault:
          type: boolean
          default: false
    SmtpConnectionUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        tls:
          type: boolean
        isDefault:
          type: boolean
    SmtpTestResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: ["integer", "null"]

    # ── API Keys ───────────────────────────────────────────────────────────
...