# lib/api-zod/src/generated/types/organizationInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CampaignForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrganizationInput {
  /** @minLength 1 */
  name: string;
  slug?: string;
  logoUrl?: string;
}